+++ interfaces/external-modules/gray-matter.d.js
@@ -1,0 +1,9 @@
+type GrayMatterResult = {
+  orig: string,
+  data: Object,
+  content: string,
+}
+
+declare module "gray-matter" {
+  declare var exports: (string: string) => GrayMatterResult
+}