+++ src/content-loader-plugins/add-body-property-from-content/__tests__/index.js
@@ -1,0 +1,17 @@
+import test from "ava"
+
+import plugin from ".."
+
+test("should add body property from content", (t) => {
+  t.deepEqual(
+    plugin({
+      frontMatter: {
+        content: "foo",
+      },
+      result: {},
+    }),
+    {
+      body: "foo",
+    }
+  )
+})