+++ src/content-loader-plugins/add-head-property-from-config/__tests__/index.js
@@ -1,0 +1,21 @@
+import test from "ava"
+
+import plugin from ".."
+
+test("should add head property from config", (t) => {
+  t.deepEqual(
+    plugin({
+      result: {},
+      options: {
+        defaultHead: {
+          "te": "st",
+        },
+      },
+    }),
+    {
+      head: {
+        "te": "st",
+      },
+    }
+  )
+})