+++ src/content-loader-plugins/add-raw-property-from-content/index.js
@@ -1,0 +1,13 @@
+// @flow
+
+export default (
+  {
+    result,
+    frontMatter,
+  }: PhenomicContentLoaderPluginInput
+): PhenomicCollectionItem => {
+  return {
+    ...result,
+    raw: frontMatter.orig,
+  }
+}