+++ src/content-loader-plugins/transform-md-body-property-to-html/index.js
@@ -4,8 +4,8 @@ import autoLinkHeadings from "remark-autolink-headings"
 import highlight from "remark-highlight.js"
 import html from "remark-html"
 
-export default (text) => (
-  remark()
+function mdify(text) {
+  return remark()
     // https://github.com/wooorm/remark-slug
     .use(slug)
 
@@ -28,4 +28,15 @@ export default (text) => (
       commonmark: true,
     })
     .toString()
-)
+}
+
+export default (
+  {
+    result,
+  }: PhenomicContentLoaderPluginInput
+): PhenomicCollectionItem => {
+  return {
+    ...result,
+    body: mdify(result.body),
+  }
+}