+++ src/index.js
@@ -1,2 +1,6 @@
 export BodyContainer from "./components/BodyContainer"
 export joinUri from "url-join"
+
+export const contentLoader = "phenomic/lib/content-loader"
+export contentLoaderDefaultPlugins from "./content-loader/default-plugins"
+export contentLoaderPlugins from "./content-loader-plugins"